Add Folders for Group, Zone, User
Code for individual features such as Group, Zone, and User will have
their own folder. So everything with Groups (GroupDAO, GroupService,
GetGroupsHandler, GetGroupsRequest/Response) will be in a single folder.
This is opposed to the method of storing files horizontally (all DAOs in
a folder, all Services in a folder, all Handlers in a folder). I chose
to do it this way because it makes more sense with the way the work is
divided. Also, if you are working on Groups, it makes sense to have
everything about a Group in one place. If I'm working on a GroupDAO, it
means I'll be likely to work on the GroupService and GroupHandler as
well. On the other hand, if I'm working on a GroupDAO, very rarely does
that mean I'll also need to work on the StatusDAO, UserDAO, and ZoneDAO
at the same time.