Shorten json for arduino
Ok guys, so the arduino has super limited memory, and the responses the
server was giving back were way too verbose for it to handle. So I made
a GetSmallStatusesResponse and a GetSmallWhereaboutsResponse that
returns the bare minimum in abbreviated json so that the arduino could
handle the response. And it works!