Fix PWM conflict with ESP8266 and Servos
Apparently, when the arduino board would access the ESP8266 wifi module,
it would mess with the PWM of all the pins. That caused the servos to go
crazy any time a network call was made. The solution was to temporarily
disconnect the servos every time we needed to talk to the server.