Improve Clock Hand Transitioning
I had some bugs in my previous code for transitioning clock hands.
Now there is the TransitionHandTo() function in Arduino that will
smoothly move a clock hand from its current position to a new position
at the rate specified in the constants inside ClockHands. Basically, you
change the value for how many milliseconds it takes to do a full circle.
I have it take 15 seconds for a full circle because I want a transition
between states to be fairly dramatic.