Remove Arduino Statuses
Turns out that the arduino board is SOO limited with memory that it
doesn't even have enough space to receive 8 full JSON statuses when
initializing. So, I simplified the clock so that it doesn't receive
statuses at all. Instead, the whereabouts that it receives contains both
the clockHandIndex and the clockHandAngle that the hand should go to.
This greatly simplifies the arduino code, which is really needed. It
also reduces the size of JSON that needs to be sent to the arduino.