Handle Errors Correctly in UserRepository
When a network request fails, it throws a ServerException. These
exceptions were not being handled correctly by the UserRepository, so
the app was crashing for simple cases like when you enter the wrong
username. I am now handling the exceptions...

Also, the logic was incorrect for logging in... The UserRepository was
returning true when an error occurred, but the LoginViewModel was
interpretting that as "true, the operation was successful"....